Add rolls method, specs pass
As discussed in the design section, we suspected we'd have a rolls method that
accepts the number of pins knocked down (the pinfall) in that roll.
But we defer writing the method until we have a failing test to prove
it is needed.

Experienced developers may scoff at this, but the point of this kata is
to get the feeling of strict adherence to the nitty-gritty, TDD.